fix(prettier): add prettylint to correct check
By default prettier check don't show any information about the
lint errors/warnings, not even if you allow debug mode.
So it is necessary to use prettylint to display which lines have mistakes.

It is typical node world problem, if you have a problem with a package,
you need to install an other package. :rofl: